requirements.txt
alabaster==0.7.16
anyio==4.4.0
asgiref==3.8.1
asttokens==2.4.1
async-property==0.2.2
Babel==2.15.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cmdstanpy==1.2.3
comm==0.2.2
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
deprecation==2.1.0
Django==5.0.6
docutils==0.21.2
emoji==2.12.1
et-xmlfile==1.1.0
Events==0.5
executing==2.0.1
Flask==3.0.3
fonttools==4.53.0
google-api-core==2.19.0
google-api-python-client==2.133.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.1
h11==0.14.0
holidays==0.50
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
image==1.5.33
imagesize==1.4.1
importlib_resources==6.4.0
ipykernel==6.29.4
ipython==8.25.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.7.2
jwcrypto==1.5.6
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
ndjson==0.3.1
nest-asyncio==1.6.0
numpy==1.26.4
openpyxl==3.1.3
opensearch-py==2.6.0
packaging==24.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
prompt_toolkit==3.0.46
prophet==1.1.5
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
Pygments==2.18.0
pyparsing==3.1.2
pyTelegramBotAPI==4.18.1
python-dateutil==2.9.0.post0
python-keycloak==4.1.0
pytz==2024.1
pyzmq==26.0.3
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.5
sphinxapi-py3==2.1.11
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
sqlparse==0.5.0
stack-data==0.6.3
stanio==0.5.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.3
